+++ models/__init__.py
@@ -14,6 +14,7 @@ from .maskformer import DinoMaskFormerHead
 from .unet import DinoUNetHead
 from .unet_lite_plus import DinoUNetLitePlusHead
 from .unet_nano import DinoUNetNanoHead
+from .unet_nano_fapm import DinoUNetNanoFAPMHead
 from .unet_v2 import DinoUNetV2Head
 from .UnetLite import DinoUNetLiteHead
 
@@ -28,7 +29,7 @@ def available_heads() -> Dict[str, HeadBuilder]:
         Dict[str, HeadBuilder]: Mapping of head names to builders.
 
     >>> sorted(available_heads().keys())
-    ['maskformer', 'unet', 'unet_lite', 'unet_lite_plus', 'unet_nano', 'unet_v2']
+    ['maskformer', 'unet', 'unet_lite', 'unet_lite_plus', 'unet_nano', 'unet_nano_fapm', 'unet_v2']
     """
 
     return {
@@ -38,6 +39,7 @@ def available_heads() -> Dict[str, HeadBuilder]:
         "unet_lite": DinoUNetLiteHead,
         "unet_lite_plus": DinoUNetLitePlusHead,
         "unet_nano": DinoUNetNanoHead,
+        "unet_nano_fapm": DinoUNetNanoFAPMHead,
     }